More tests for upgrading

+++ private/enterprise/enterprise-it/security-it/src/test/java/com/neo4j/server/security/enterprise/systemgraph/SystemGraphRealmUpgradeIT.java
@@ -19,12 +19,19 @@ import java.util.Map;
 
 import org.neo4j.cli.ExecutionContext;
 import org.neo4j.commandline.admin.security.SetDefaultAdminCommand;
+import org.neo4j.commandline.admin.security.SetInitialPasswordCommand;
 import org.neo4j.configuration.GraphDatabaseSettings;
 import org.neo4j.dbms.api.DatabaseManagementService;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Result;
 import org.neo4j.graphdb.Transaction;
+import org.neo4j.internal.kernel.api.security.AuthenticationResult;
+import org.neo4j.internal.kernel.api.security.LoginContext;
 import org.neo4j.io.fs.UncloseableDelegatingFileSystemAbstraction;
+import org.neo4j.kernel.api.security.AuthManager;
+import org.neo4j.kernel.api.security.exception.InvalidAuthTokenException;
+import org.neo4j.kernel.internal.GraphDatabaseAPI;
+import org.neo4j.server.security.auth.SecurityTestUtils;
 import org.neo4j.test.TestDatabaseManagementServiceBuilder;
 import org.neo4j.test.extension.Inject;
 import org.neo4j.test.extension.testdirectory.TestDirectoryExtension;
@@ -33,17 +40,18 @@ import org.neo4j.test.rule.TestDirectory;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.equalTo;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.hamcrest.Matchers.oneOf;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.neo4j.configuration.GraphDatabaseSettings.SYSTEM_DATABASE_NAME;
+import static org.neo4j.kernel.api.security.AuthManager.INITIAL_PASSWORD;
 import static org.neo4j.kernel.api.security.AuthManager.INITIAL_USER_NAME;
 
 @TestDirectoryExtension
 class SystemGraphRealmUpgradeIT
 {
+    private static final String DEFAULT_PASSWORD = "bar";
+
     @Inject
     private TestDirectory testDirectory;
 
@@ -58,20 +66,12 @@ class SystemGraphRealmUpgradeIT
 
         // THEN
         GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
-        try ( Transaction tx = database.beginTx() )
-        {
-            Result users = tx.execute( "SHOW USERS" );
-            assertTrue( users.hasNext() );
-            Map<String,Object> user = users.next();
-            assertThat( user.get( "user" ), equalTo( INITIAL_USER_NAME ) );
-            assertThat( (List<String>) user.get( "roles" ), containsInAnyOrder( PredefinedRoles.ADMIN ) );
-            assertFalse( users.hasNext() );
-            tx.commit();
-        }
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder( new UserResult( INITIAL_USER_NAME, List.of( PredefinedRoles.ADMIN ) ) ) );
     }
 
     @Test
-    void shouldUpgradeWithSingleRenamedUser()
+    void shouldUpgradeWithSingleCustomUser()
     {
         // GIVEN
         createCommunityWithUsers( "Alice" );
@@ -81,20 +81,12 @@ class SystemGraphRealmUpgradeIT
 
         // THEN
         GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
-        try ( Transaction tx = database.beginTx() )
-        {
-            Result users = tx.execute( "SHOW USERS" );
-            assertTrue( users.hasNext() );
-            Map<String,Object> user = users.next();
-            assertThat( user.get( "user" ), equalTo( "Alice" ) );
-            assertThat( (List<String>) user.get( "roles" ), containsInAnyOrder( PredefinedRoles.ADMIN ) );
-            assertFalse( users.hasNext() );
-            tx.commit();
-        }
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder( new UserResult( "Alice", List.of( PredefinedRoles.ADMIN ) ) ) );
     }
 
     @Test
-    void shouldUpgradeDefaultUserAndOtherUser()
+    void shouldUpgradeDefaultAndOtherUser()
     {
         // GIVEN
         createCommunityWithUsers( INITIAL_USER_NAME, "Alice" );
@@ -104,20 +96,27 @@ class SystemGraphRealmUpgradeIT
 
         // THEN
         GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
-        try ( Transaction tx = database.beginTx() )
-        {
-            List<UserResult> users = new ArrayList<>();
-            Result result = tx.execute( "SHOW USERS" );
-            while ( result.hasNext() )
-            {
-                Map<String,Object> user = result.next();
-                users.add( new UserResult( (String) user.get( "user" ), (List<String>) user.get( "roles" ) ) );
-            }
-            assertThat( users, containsInAnyOrder(
-                    new UserResult( INITIAL_USER_NAME, List.of( PredefinedRoles.ADMIN ) ),
-                    new UserResult( "Alice", Collections.emptyList() ) ) );
-            tx.commit();
-        }
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder(
+                new UserResult( INITIAL_USER_NAME, List.of( PredefinedRoles.ADMIN ) ),
+                new UserResult( "Alice", Collections.emptyList() ) )
+        );
+    }
+
+    @Test
+    void shouldChangeInitialPasswordAndSetAdmin() throws InvalidAuthTokenException
+    {
+        createCommunityWithUsers( INITIAL_USER_NAME );
+
+        // WHEN
+        setInitialPassword( "abc123" );
+        DatabaseManagementService enterpriseDbms = getEnterpriseManagementService();
+
+        // THEN
+        GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder( new UserResult( INITIAL_USER_NAME, List.of( PredefinedRoles.ADMIN ) ) ) );
+        assertLogin( (GraphDatabaseAPI) database, INITIAL_USER_NAME, "abc123" );
     }
 
     @Test
@@ -159,20 +158,94 @@ class SystemGraphRealmUpgradeIT
 
         // THEN
         GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder(
+                new UserResult( "Alice", List.of( PredefinedRoles.ADMIN ) ),
+                new UserResult( "Bob", Collections.emptyList() ) )
+        );
+    }
+
+    @Test
+    void shouldSetCustomAsAdminWhenSetAsDefaultAdmin()
+    {
+        // GIVEN
+        createCommunityWithUsers( INITIAL_USER_NAME, "Alice" );
+
+        // WHEN
+        setDefaultAdmin( "Alice" );
+        DatabaseManagementService enterpriseDbms = getEnterpriseManagementService();
+
+        // THEN
+        GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder(
+                new UserResult( "Alice", List.of( PredefinedRoles.ADMIN ) ),
+                new UserResult( INITIAL_USER_NAME, Collections.emptyList() ) )
+        );
+    }
+
+    @Test
+    void shouldSetInitialUserAsAdminWhenSetAsDefaultAdmin()
+    {
+        // GIVEN
+        createCommunityWithUsers( INITIAL_USER_NAME, "Alice" );
+
+        // WHEN
+        setDefaultAdmin( INITIAL_USER_NAME );
+        DatabaseManagementService enterpriseDbms = getEnterpriseManagementService();
+
+        // THEN
+        GraphDatabaseService database = enterpriseDbms.database( SYSTEM_DATABASE_NAME );
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder(
+                new UserResult( INITIAL_USER_NAME, List.of( PredefinedRoles.ADMIN ) ),
+                new UserResult( "Alice", Collections.emptyList() ) )
+        );
+    }
+
+    @Test
+    void shouldSetCustomAsAdminAndInitialPasswordForInitialUser() throws InvalidAuthTokenException
+    {
+        // GIVEN
+        createCommunityWithUsers( INITIAL_USER_NAME, "Alice" );
+
+        // WHEN
+        setDefaultAdmin( "Alice" );
+        setInitialPassword( "foo" );
+        DatabaseManagementService enterpriseDbms = getEnterpriseManagementService();
+
+        // THEN
+        GraphDatabaseAPI database = (GraphDatabaseAPI) enterpriseDbms.database( SYSTEM_DATABASE_NAME );
+        List<UserResult> users = getUsers( database );
+        assertThat( users, containsInAnyOrder(
+                new UserResult( "Alice", List.of( PredefinedRoles.ADMIN ) ),
+                new UserResult( INITIAL_USER_NAME, Collections.emptyList() ) )
+        );
+        assertLogin( database, "Alice", DEFAULT_PASSWORD );
+        assertLogin( database, INITIAL_USER_NAME, INITIAL_PASSWORD );
+    }
+
+    private void assertLogin( GraphDatabaseAPI database, String username, String password ) throws InvalidAuthTokenException
+    {
+        AuthManager authManager = database.getDependencyResolver().resolveDependency( AuthManager.class );
+        LoginContext loginContext = authManager.login( SecurityTestUtils.authToken( username, password ) );
+        assertThat( loginContext.subject().getAuthenticationResult(), oneOf( AuthenticationResult.SUCCESS, AuthenticationResult.PASSWORD_CHANGE_REQUIRED ) );
+    }
+
+    private List<UserResult> getUsers( GraphDatabaseService database )
+    {
+        List<UserResult> users = new ArrayList<>();
         try ( Transaction tx = database.beginTx() )
         {
-            List<UserResult> users = new ArrayList<>();
             Result result = tx.execute( "SHOW USERS" );
             while ( result.hasNext() )
             {
                 Map<String,Object> user = result.next();
                 users.add( new UserResult( (String) user.get( "user" ), (List<String>) user.get( "roles" ) ) );
             }
-            assertThat( users, containsInAnyOrder(
-                    new UserResult( "Alice", List.of( PredefinedRoles.ADMIN ) ),
-                    new UserResult( "Bob", Collections.emptyList() ) ) );
             tx.commit();
         }
+        return users;
     }
 
     private static class UserResult
@@ -220,10 +293,21 @@ class SystemGraphRealmUpgradeIT
         GraphDatabaseService database = managementService.database( SYSTEM_DATABASE_NAME );
         try ( Transaction tx = database.beginTx() )
         {
-            tx.execute( String.format( "DROP USER %s", INITIAL_USER_NAME ) ).close();
+            boolean dropInitialUser = true;
             for ( String user : users )
             {
-                tx.execute( String.format( "CREATE USER %s SET PASSWORD 'bar' CHANGE NOT REQUIRED", user ) ).close();
+                if ( user.equals( INITIAL_USER_NAME ) )
+                {
+                    dropInitialUser = false;
+                }
+                else
+                {
+                    tx.execute( String.format( "CREATE USER %s SET PASSWORD '%s' CHANGE NOT REQUIRED", user, DEFAULT_PASSWORD ) ).close();
+                }
+            }
+            if ( dropInitialUser )
+            {
+                tx.execute( String.format( "DROP USER %s", INITIAL_USER_NAME ) ).close();
             }
             tx.commit();
         }
@@ -239,6 +323,18 @@ class SystemGraphRealmUpgradeIT
                                       mock( PrintStream.class ),
                                       testDirectory.getFileSystem() ) );
         CommandLine.populateCommand( command, username );
+        command.execute();
+    }
+
+    private void setInitialPassword( String password )
+    {
+        final var command = new SetInitialPasswordCommand(
+                new ExecutionContext( testDirectory.homeDir().toPath(),
+                                      new File( testDirectory.homeDir(), "conf" ).toPath(),
+                                      mock( PrintStream.class ),
+                                      mock( PrintStream.class ),
+                                      testDirectory.getFileSystem() ) );
+        CommandLine.populateCommand( command, password );
         command.execute();
     }
 }
